+++ services/UserService.js
@@ -8,24 +8,21 @@ const UserService = {
     login: async (req, res) => {
       try {
         const user = await User.findOne({ email: req.body.email })
-        console.log(user)
         if(!user)
           res.status(404).send("User not registered")
         else {
           const pass = await bcrypt.compare(req.body.password, user.password)
-          console.log(pass)
           if(!pass)
             res.status(401).send("Wrong password")
           else {
             const sess = await Session.findOne({ user_id: user.id })
-            console.log(sess)
             if(sess)
               res.status(401).send("User already logged in")
             else {
               const token = crypto.randomBytes(64).toString('hex');
               const session = Session.create({
                 token: token,
-                ser_id: user.id
+                user_id: user.id
               })
               res.status(200).json({ user, token })
             }
@@ -52,13 +49,13 @@ const UserService = {
       }
     },
     getUserById: async (req, res) => {
-      const token = req.headers['auth-token'];
-      let session = await Session.findOne({token}).lean().exec();
-      console.log(session)
-      if (!session) {
-        res.status(404).json();
-      } else {
-        try {
+      try {
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else {
           const user = await User.findOne( { id: req.params.id.toString }).lean().exec()
           console.log(user)
           res.status(200).json({
@@ -70,71 +67,63 @@ const UserService = {
             company: user.company,
             isBuyer: user.isBuyer
           })
-        } catch (error) {
-          res.status(400).send()        
         }
+      } catch (error) {
+        res.status(400).send()        
       }
     },
     getUsers: async (req, res ) => {
-      const token = req.headers['auth-token'];
-      let session = await Session.findOne({token}).lean().exec();
-      console.log(token)
-      if (!session) {
-        res.status(404).json();
-      } 
-      else 
       try {
-        const page = req.query.page;
-        const size = req.query.size;
-        const users = await User.find({}).skip(parseInt(size * page)).limit(parseInt(size)).lean().exec();
-        res.json(users.map(function(user) { 
-          return { 
-            id: user.id,
-            name: user.name,
-            phone: user.phone,
-            email: user.email,              
-            address: user.address,
-            company: user.company,
-            isBuyer: user.isBuyer 
-          }; 
-        }))
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else {
+          const page = req.query.page;
+          const size = req.query.size;
+          const users = await User.find({}).skip(parseInt(size * page)).limit(parseInt(size)).lean().exec();
+          res.json(users.map(function(user) { 
+            return { 
+              id: user.id,
+              name: user.name,
+              phone: user.phone,
+              email: user.email,              
+              address: user.address,
+              company: user.company,
+              isBuyer: user.isBuyer 
+            }; 
+          }))
+        }
       } catch (err) {
         res.status(400).json({ message: 'There has been an error' })
       }
     },
     getBuyers: async (req, res) => {
-      const token = req.headers['auth-token'];
-      let session = await Session.findOne({token}).lean().exec();
-      console.log(token)
-      if (!session) {
-        res.status(404).json();
-      } 
-      // else if(user.isBuyer) {
-      //   res.status(401).send("Unauthorized to see other buyers")
-      // }
-      else {
-        try {      
+      try {
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else {
           const page = req.query.page;
           const size = req.query.size;
           const users = await User.find({ isBuyer: true }).skip(parseInt(size * page)).limit(parseInt(size)).lean().exec();
           res.json(users)
-        } catch (err) {
-          res.status(400).json({ message: 'There has been an error' })
         }
-      } 
+      } catch (err) {
+        res.status(400).json({ message: 'There has been an error' })
+      }
     },
     getSellers: async (req, res) => {
-      const token = req.headers['auth-token'];
-      let session = await Session.findOne({token}).lean().exec();
-      console.log(token)
-      if (!session) {
-        res.status(404).json();
-      } 
-      // else if (!user.isBuyer) {
-      //   res.status(401).send("Unauthorized to see other sellers")
-      // }
-      else {
-        try {          
+      try {
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else {         
           const page = req.query.page;
           const size = req.query.size;
           const users = await User.find({ isBuyer: false }).skip(parseInt(size * page)).limit(parseInt(size)).lean().exec();
@@ -149,9 +138,9 @@ const UserService = {
               isBuyer: user.isBuyer 
             }; 
           }))
-        } catch (err) {
-          res.status(400).json({ message: 'There has been an error' })
         }
+      } catch (err) {
+        res.status(400).json({ message: 'There has been an error' })
       } 
     },
     createUser: async (req, res) => {
@@ -175,8 +164,15 @@ const UserService = {
     },
     updateUser: async (req, res) => {
       try {
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else {  
           const user = await User.findByIdAndUpdate(req.params.id)
-          res.send({ user })
+          res.status(200).send({user})
+        }
       } catch (err) {
           console.log({ err })
           res.status(400).send({ message: 'error' })
@@ -184,8 +180,15 @@ const UserService = {
     },
     deleteUser: async (req, res) => {
       try {
+        const token = req.headers['auth-token'];
+        let session = await Session.findOne({token}).lean().exec();
+        if (!session) {
+          res.status(401).json("Session token is not valid");
+        } 
+        else { 
           const result = await User.findByIdAndDelete(req.params.id)
-          res.send({ result })
+          res.status(200).send({ result })
+        }
       } catch(err) {
           console.log({ err })
           res.status(400).send({ message: 'error' })
